powerfit-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
python-dotenv==1.0.0
bcrypt==4.0.1
pymysql==1.1.2
stripe==7.5.0
email-validator==2.1.1